backend/requirements.txt
django~=3.0.0
django-debug-toolbar
pillow
djangorestframework
djangorestframework-jwt